Fix the type of the 4th arguement of getgrnam_r().
(backport <https://github.com/rust-lang/libc/pull/4656>)
(cherry picked from commit 8236b564f86744ec040343704e75af814d6244ad)

+++ src/unix/aix/mod.rs
@@ -2927,7 +2927,7 @@ extern "C" {
         name: *const c_char,
         grp: *mut crate::group,
         buf: *mut c_char,
-        buflen: c_int,
+        buflen: size_t,
         result: *mut *mut crate::group,
     ) -> c_int;
     pub fn getgrset(user: *const c_char) -> *mut c_char;
